Updated translation
translated time to "Time to discharge" with "verbleibend" as in "remaining", which is more intuitive.
Translating it literally to time the battery takes to discharge does not consider how heavy the usage is on it. Translating it to remaining, displays what this tool intended to show to the user - how long will the battery hold up, considering the usage right now.